+++ src/libcsg/modules/io/dlpolytopologyreader.cc
@@ -110,9 +110,8 @@ bool DLPOLYTopologyReader::_isKeyInt(const string &line, const char *wspace,
 {
   // split directives consisting of a few words: the sought keyword must be the
   // first one!
-  Tokenizer tok(line, wspace);
-  vector<string> fields;
-  tok.ToVector(fields);
+
+  vector<string> fields = Tokenizer(line, wspace).ToVector();
 
   ival = 0;
 
@@ -243,9 +242,7 @@ bool DLPOLYTopologyReader::ReadTopology(string file, Topology &top) {
         line = " ";
         sl >> line;  // rest of the atom line
 
-        Tokenizer tok(line, WhiteSpace);
-        vector<string> fields;
-        tok.ToVector(fields);
+        vector<string> fields = Tokenizer(line, WhiteSpace).ToVector();
 
 #ifdef DEBUG
         cout << "Rest atom specs from dlpoly topology : '" << line << "'"